Add the original example to the .sln and update with current API
