+++ example/Example/k8s/collector-config.yaml
