+++ example/Example/k8s/collector-pod.yaml
